Error handle pwdCommand() and make execCommand() more readable
The case where getenv() returns NULL in pwdCommand() was not handled.

Finally, the nested ifs at the beginning of execCommand() made the function
harder to read. Also the cases 0 and 1 for argc were handled impicitly which
obscures the code and potentially confuses the reader. In addition to all of
that, the execution and error handling of execvp() seem to make execCommand()
more bloated than it needs to be and also distracts away from the main logic
of the function. Therefore I tried to solve these by explicitely handling
the 0 and 1 cases for argc, and introducing a new function run_execvp() to
handle the execution and error handling of execvp().