Implement color feature of shell and un-const a parameter
Now colors can easily be added to the prompts of the shell.Also Implemented a
random color generator function that randomly chooses between 6 colors each
time that is called.

The last paramter of run_execvp() was a const pointer. The compiler warned
me when I tried to compile the program, since the function tries to free it.
But this is not possible as it is a const variable and so I changed it to a
normal pointer.