Put headers in source files,declare function in related header and pass argc and argv to functions
Unfortunately not being aware of how one should put #includes in
files, I always put all of them in the respective header files. That
is wrong. One should only put #include in headers if there is any need
for it in the declaration of variables of functions. Otherwise one should
put them in the source files.

read_storeCommand() was previously declared
and defined in basic.h and basic.c. The problem is that the function had
2 jobs which goes against the single responsibility principle for functions.
It also depended on add() from history.h, which also shows that given its
job, its place is in history.h not basic.h. So I defined storeCommand() in
history.c and deleted read_storeCommand() entirely from basic files.

Finally, the program was designed as to make argc and argv global variables
in the project. That had the benefit of not repeatedly bloating function
declarations with extra parameters. But unfortunately this approach does
introduce several problems like not being thread safety, harder to maintain
and find bugs etc. So I decided to pass them to functions and make them local
rather global variables in the project.