Make all of the variables in history.h global
freeAllBlock(), and several of the variables in history.h are used in
multiple places. In order to call the function successfully, we have to
make all of the variables in history.h global and share it between
the files that need it. So I declared them extern in .h file and defined
them in history.c file. Any file that needs access to them can then just
include the .h file.