Change return type of findIndexOfLastSlash()
The function tries to first find how many slashes there are in the currentPath.
If there were none then it returns 0. If not, it goes through currentPath to
find the index of the last one. In case indexOfLastSlash tries to go beyond
the size of currentPath, we need to somehow return and inform the caller of
such error. The previous return type, size_t was not appropriate for such task.
Since any return type is 0 or more and it is not possible to differentiate
whether this is an appropriate index of the array or meant to be an error.
So I changed it to int in order to return -1 in such situation.