Transfer commandToStoreInHistBlock() from basic.h to history.h
This function's responsibility is to take argv and traverse it and then give us
the command, where each token is in command seperated by spaces. The reason for its
existence is that argv members point to inputWords[] and each contiguous block of
non-null chars in inputWords[] is seperated by \0 chars. So each time argv is updated
we need to traverse it and get the input command in its "natural" format where each
word is seperated by just a space. Such transfer is needed as because based on how
struct CommandBlock are defined, they need to point to a command in its natural format,
which only this function can provide.