Fix function declarations in header and eliminate local variable countList
Since countList became global variable, there is no need for local variable anymore.
On top of that, introducing local one causes a myriads of bugs, all of which stem from
the fact that by introducing a local variable as countList we do not reset countList back to
0. This causes our next input to increment countList starting from the previous value that it
held from the previous command. If that value is 0 then everything will be fine. If not, then
it will count more lists that there are and we will most likely get a segmentation fault as there
are not that many commandBlocks for initializeArrayOfCommandBlocks() to initialize in the first place.
So the solution was to just make sure GlobalSpecialCharCounter() resets the global variable countList
to 0 each time a new input is entered.