Add source and header files for weather.py
It is possible to execute weather.py without much hassle. However
in order to execute it, the user is forced to type a long command
which defeats its purpose: to inquiry about the weather as fast as
possible. So in order to solve this issue, I am introducing a weather
function which significantly shortens the length of the command that
the user has to type.