Remove signal handler for SIGINT
The reason I included a signal handler for SIGINT was to make
sure that all the history blocks are freed when a SIGINT signal
is sent. But I have come to realize that the memory leaked due to
SIGINT is not important and it is better to leave the clean up to
the OS. Another important is that having exit() in a signal handler
is undefined behaviour and so it is best to just not handle the SIGINT
signal. Therefore I removed SIGINT case from the signal handler.