Eliminate redundant printShellSign() function
The printing of the shell sign "FAN" along with the current directory was
printShellSign() responsibility. However after the input was handled in the
Terminal instance and the input was transferred to inputWords[], the start_shell()
is called which causes the shell sign along with the current directory to be printed
again. Therefore to fix this issue, I transferred the responsibility of printing
the shell sign to the Terminal instance that gets instantiated at the beginning of
main(). As a result, printShellSIgn() is no longer needed.