Introduce a new function that seperately handles the disabling of raw mode
Raw mode could only have been disabled by destroying the Terminal object. This
did indeed cause a few problems. For example, if we wanted to get the result of
the typed command, we had to handle the cursor's position at the same time that
the result was being printed to the screen. This is not easily achievable simply
because the Terminal object's cursor is not synced with the programs that will be
run when we type the command. I have decided to switch to canonical mode
when the input is transferred to inputWords[]. As a result, I had to introduce
a function that handles this task instead of the destructor itself.