Enable destructor to disable raw mode before destroying instance and throw string instead of char*
When Ctrl-q was pressed exception was thrown. But the exception was in
const char*. So I had to initialize a string instance before throw. However
this introduced another problem. The destructor destroys the Shell_Terminal
object without disabling raw mode. This will end the program, leaving the shell
in raw mode. So in order to fix this I made sure destructor disables raw mode
before destroying the instance.