Fix a bug in printInputAfterDeletion()
Whenever the cursor was at the beginning of a line and
the user pressed backspace, cursorX was decremented as well.
This was caused by updateCursorPos() being used right after the
for loop. This decrementation caused the cursor to be on the line
above the line that it was supposed to be. So in order to avoid this
problem, I manually handled the case when curseY == 0 when the for
loop is finished. The rest of the changes are rather trivial and do
not need comments.