Keep track of cursorX before erasing input when backspace is pressed
cursorX had to be kept track of before we restart its value
alongside cursorY. Otherwise we would have trouble determining
the position of the cursor in printInputAfterDeletion() after
we have gone through the for loop. Therefore we record its value
in a dummy value, just like cursorY, and then take care of the
edge case when cursorY == 0.