Define a function to transfer the history command to input list
When the user presses the up arrow key, the last command that was
inserted by them is shown. However they cannot modify it, because
the cursor is controlled by the iterator of input list. In order to
be able to modify it; we ought to fill the input list with the previous
command characters and then we can move around the cursor freely and
modify the previous command.