Implement going to the end of the command by pageDownAction() and fix bugs
Going to the end of the typed command is implemented by pageDownAction().
When pg dn is pressed we get \1xb[6~ , and each one of these characters
that make it up are processed by their respective ascii code. So since
the ascii code of '6' is 54, we should include 54 in the case of switch
and not 6 which is an ascii code for another character altogether. The
same exact reasoning led me to change case 5 to case 53 in the switch of
checkForArrowKeys().