Change the type that list container historyCommands can contain
It seemed ineffecient that historyCommands contained the whole historyBlock
and only used its command field. So the solution I came up with was to change
the data type that historyCommands could hold and the best candidate for it
was char*. After changing this data type, I edited the parts of the code that
used this list in order to accommidate for the mentioned change.