Refactor paramters of functions && use execlp in 1st if of run_execvp()
argv and argc are shared global variables, so there is no need to pass them
to functions. Therefore I refactored the function delarations that used them.
execvp() was used when newArgv is NULL, which is a mistake since the second argument
of execvp() should not be NULL, so I used execlp instead.