Fix cdCommand() bug
I tried to eliminate the need for argc and argv being passed to functions
by paramters, since they are global variables that are shared between files.
However the cdCommand()'s argc parameter name before tricked me into thinking
that it was the global argc. But I was wrong. In this function argc is solely
used to find out how many strings come after cd command that it typed by the
user. So cd could be used in a list of && and come for example as the last
command in the list and argc in this context would mean how many strings
, cd also counted, are there in the cd command and not how many strings there
are in total in the input that the user has typed in. So I re-added the second
parameter of cdCommand() and renamed it to something less deceiving.