Change the name of a misleading parameter name in both functions
argc is not the value that is used in these 2 functions. These 2 functions
require the number of strings in a single command that itself is part of a
command block. argc is the value of the whole input entered by the user, which
if it contitutes of a single command, will be the same as that of the number of
strings in a command. Otherwise, these 2 values are different and so I changed
the parameters name to represent its actual purpose.